split test cross entropy loss
